skills/manual-to-ppt/requirements.txt
openpyxl>=3.1.0
python-pptx>=1.0.0
Pillow>=10.0.0

